Fix code compliance
